pyproject.toml
[project]
name = "deepvision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepface>=0.0.95",
    "dotenv>=0.9.9",
    "face-comparer>=2.0",
    "face-recognition>=1.3.0",
    "flask>=3.1.2",
    "flask-cors>=4.0.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "supabase>=2.24.0",
    "tf-keras>=2.20.1",
    "typing>=3.10.0.0",
    "requests>=2.32.3",
    "pyperclip>=1.11.0",
    "transformers>=4.57.1",
    "torch>=2.9.1",
]
